Fixed hashing function
So when I developed the group messaging app and the login app on my
Linux OS it prefered the hash encryption method for password hashes,
owing to the md5 function being less secure. However, when attempting to
use it on some old versions on windows, php didn't like this method
without more information, which couldn't be required. So I just used
md5, which should be sufficent enough.